chore: more accurate description of what component actually is
